requirements.txt
Flask==2.3.2
ultralytics==8.0.0
opencv-python==4.8.0.76
numpy==1.24.3
pyserial==3.5
